% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddxGenxA.R
\name{AddxGenxA}
\alias{AddxGenxA}
\title{Add xG and xA}
\usage{
AddxGenxA(df)
}
\arguments{
\item{df}{A dataframe of the game.}
}
\value{
data frame with the values.
}
\description{
Add xG and xA
}
\examples{
AddxGenxA(df)
}
