% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateChance.R
\name{calculateChance}
\alias{calculateChance}
\title{Calculate winning chances}
\usage{
calculateChance(team1, team2, p)
}
\arguments{
\item{team1}{Home team xg vector}

\item{team2}{Away team xg vector}

\item{p}{Number of sims}
}
\value{
chance
}
\description{
Calculate winning chances
}
\examples{
calculateChance(team1,team2,p)
}
