% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateChance.R
\name{calculateWinner}
\alias{calculateWinner}
\title{Needed for CalculateChance()}
\usage{
calculateWinner(home, away)
}
\arguments{
\item{home}{home team xg}

\item{away}{away team xg}
}
\value{
winner
}
\description{
Needed for CalculateChance()
}
\examples{
calculateWinner(home, away)
}
