% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertIDOneGame.R
\name{ConvertIDOneGame}
\alias{ConvertIDOneGame}
\title{change PlayerId, TeamId and relevantPlayerId to names}
\usage{
ConvertIDOneGame(DATA, OneGamePlayers)
}
\arguments{
\item{DATA}{A data frame}

\item{OneGamePlayers}{Date frame with player names and id's.}
}
\value{
\code{DATA} with the names in it
}
\description{
change PlayerId, TeamId and relevantPlayerId to names
}
\examples{
ConvertIDOneGame(DATA,OneGamePlayers)
}
