% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(x_pos, y_pos)
}
\arguments{
\item{x_pos}{x}

\item{y_pos}{y}
}
\value{
distance in meters.
}
\description{
Function that calculates distance to goal
}
\examples{
distance(10,10)
}
