% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{distanceCarry}
\alias{distanceCarry}
\title{distanceCarry}
\usage{
distanceCarry(x_start, y_start, x_end, y_end)
}
\arguments{
\item{x_start}{x}

\item{y_start}{y}

\item{x_end}{endX}

\item{y_end}{endY}
}
\value{
distance in meters.
}
\description{
Function that calculates distance of carry
}
\examples{
distanceCarry(75,10,75,5)
}
