% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{distanceToGoal}
\alias{distanceToGoal}
\title{distanceToGoal}
\usage{
distanceToGoal(x_pos, y_pos)
}
\arguments{
\item{x_pos}{x}

\item{y_pos}{y}
}
\value{
distance in meters.
}
\description{
Function that calculates distance to goal
}
\examples{
distanceToGoal(10,10)
}
