% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEverything.R
\name{doEverything}
\alias{doEverything}
\title{Make the ready for dashboard}
\usage{
doEverything(json_file, HomeTeam, AwayTeam)
}
\arguments{
\item{jsonfile}{path to json file of match}
}
\value{
png of dashboard
}
\description{
Make the ready for dashboard
}
\examples{
doEverything(json_file, HomeTeam, AwayTeam)
}
