% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossesKKD.R
\name{getCrossesKKD}
\alias{getCrossesKKD}
\title{Function that adds cross = 1 or cross = 0
for xT}
\usage{
getCrossesKKD(df)
}
\arguments{
\item{df}{event data}
}
\description{
Function that adds cross = 1 or cross = 0
for xT
}
\examples{
getCrosses(df)
}
\keyword{crosses}
