% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHashtag.R
\name{GetHashtag}
\alias{GetHashtag}
\title{Add hashtag for twitter}
\usage{
GetHashtag(Team)
}
\arguments{
\item{team}{A team of the game.}
}
\value{
hastag.
}
\description{
Add hashtag for twitter
}
\examples{
GetHashtag(Team)
}
