% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelatedKKD.R
\name{getRelatedKKD}
\alias{getRelatedKKD}
\title{Get related player for shots}
\usage{
getRelatedKKD(With_names, df)
}
\arguments{
\item{df}{A data frame}
}
\value{
df with relatedPlayerId added
}
\description{
Get related player for shots
}
\examples{
getRelatedKKD(dfcsv)
}
