% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateChance.R
\name{GetWedstrijden}
\alias{GetWedstrijden}
\title{Needed for CalculateChance()}
\usage{
GetWedstrijden(Wedstrijd)
}
\arguments{
\item{wedstrijd}{data frame}
}
\value{
kansen
}
\description{
Needed for CalculateChance()
}
\examples{
GetWedstrijden(wedstrijd)
}
