% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{goal_angle}
\alias{goal_angle}
\title{goal_angle}
\usage{
goal_angle(x_pos, y_pos)
}
\arguments{
\item{x_pos}{x}

\item{y_pos}{y}
}
\value{
angle in degrees
}
\description{
Function that calculates angle to goal
}
\examples{
goal_angle(10,10)
}
