% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePartsPatchwork.R
\name{MakePartsPatchwork}
\alias{MakePartsPatchwork}
\title{Make the dashboard}
\usage{
MakePartsPatchwork(Wedstrijd)
}
\arguments{
\item{Wedstrijd}{A game with xT and xG added.}
}
\value{
png of dashboard
}
\description{
Make the dashboard
}
\examples{
MakePartsPatchwork(Wedstrijd)
}
