% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePartsPatchworkKKD.R
\name{MakePartsPatchworkKKD}
\alias{MakePartsPatchworkKKD}
\title{Make the dashboard}
\usage{
MakePartsPatchworkKKD(listEntry1, listEntry2, listEntry3)
}
\arguments{
\item{Wedstrijd}{A game with xT and xG added.}
}
\value{
png of dashboard
}
\description{
Make the dashboard
}
\examples{
MakePartsPatchwork(Wedstrijd)
}
