% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minutes.R
\name{Minutes}
\alias{Minutes}
\title{Get minutes played}
\usage{
Minutes(dfcsv)
}
\arguments{
\item{dfcsv}{A data frame}
}
\value{
summarise of players and minutes, added red card
}
\description{
Get minutes played
}
\examples{
Minutes(dfcsv)
}
