% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinutesKKD.R
\name{MinutesKKD}
\alias{MinutesKKD}
\title{Get minutes played}
\usage{
MinutesKKD(df_rel)
}
\arguments{
\item{dfcsv}{A data frame}
}
\value{
summarise of players and minutes, added red card
}
\description{
Get minutes played
}
\examples{
Minutes(dfcsv)
}
