% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictxG.R
\name{PredictxG}
\alias{PredictxG}
\title{Function that adds xG for the shots
reads the model file first}
\usage{
PredictxG(df_for_pred)
}
\arguments{
\item{df}{shot data}
}
\description{
Function that adds xG for the shots
reads the model file first
}
\examples{
PredictxG(df)
}
\keyword{predict}
