% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepAGame.R
\name{PrepAGame}
\alias{PrepAGame}
\title{Function that prepares a dataframe for xG
makes a shotmap of shots against Eredivisie Team}
\usage{
PrepAGame(df)
}
\arguments{
\item{df}{event data}
}
\description{
Function that prepares a dataframe for xG
makes a shotmap of shots against Eredivisie Team
}
\examples{
PrepAGame2.0(df)
}
\keyword{shotmap}
