% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepAGameKKD.R
\name{PrepAGameKKD}
\alias{PrepAGameKKD}
\title{Function that prepares a dataframe for xG
makes a shotmap of shots against Eredivisie Team}
\usage{
PrepAGameKKD(df_rel)
}
\arguments{
\item{df}{event data}
}
\description{
Function that prepares a dataframe for xG
makes a shotmap of shots against Eredivisie Team
}
\examples{
PrepAGameKKD(df)
}
\keyword{shotmap}
