% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapeGame.R
\name{ScrapeGame}
\alias{ScrapeGame}
\title{Scrape a game}
\usage{
ScrapeGame(url)
}
\arguments{
\item{url}{A url from whoscored}
}
\value{
df of game
}
\description{
Scrape a game
}
\examples{
ScrapeGame(url)
}
