% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateChance.R
\name{testShots}
\alias{testShots}
\title{Needed for CalculateChance()}
\usage{
testShots(XG)
}
\arguments{
\item{xG}{xG to sim}
}
\value{
Number of goals
}
\description{
Needed for CalculateChance()
}
\examples{
testShots(xG)
}
