% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TweetTweet.R
\name{TweetTweet}
\alias{TweetTweet}
\title{Add together two numbers}
\usage{
TweetTweet(Wedstrijd = data.frame(1, 1), number)
}
\arguments{
\item{Wedstrijd}{A df from the game or empty}

\item{number}{what do we tweet?.}
}
\value{
it tweets!
}
\description{
Add together two numbers
}
\examples{
TweetTweet(Wedstrijd, 1)
1 = game
2 = xRank
3 = topscorers
}
