% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeOfPlay.R
\name{typeOfPlay}
\alias{typeOfPlay}
\title{Add type of play}
\usage{
typeOfPlay(OneGame)
}
\arguments{
\item{OneGame}{A game.}
}
\value{
df of game with Typ_of_play column
}
\description{
Add type of play
}
\examples{
typeOfPlay(OneGame)
}
